Initial port of the color and indexer classes to ES6 Modules.
